+++ pkg/service/service.go
@@ -3,5 +3,6 @@ package service
 import "math/big"
 
 type Service interface {
-	Execute(int, int) ([]*big.Int, error)
+	//Execute – исполнитель
+	Execute(int64, int64) ([]*big.Int, error)
 }
